backend/requirements.txt
selenium
webdriver-manager
selenium-stealth
chromedriver-autoinstaller
pydantic>=2.0.0
fastapi
uvicorn
google-generativeai
python-dotenv
redis
